reafctoring: move output view to view folder and separate output view logics in a new controller and model.
